backend/requirements.txt
flask==3.1.0
flask-cors==4.0.0
python-dotenv==1.0.1
stravalib==2.1.0
pandas==2.2.3
requests==2.32.3 